+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/ui/ImageComposables.kt
@@ -1,15 +1,11 @@
 package com.thekeeperofpie.artistalleydatabase.anime.ui
 
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ImageNotSupported
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
 import androidx.compose.ui.draw.blur
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.rememberVectorPainter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
@@ -17,13 +13,55 @@ import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import coil.compose.AsyncImagePainter
 import coil.request.ImageRequest
-import com.google.accompanist.placeholder.PlaceholderHighlight
-import com.google.accompanist.placeholder.material.placeholder
-import com.google.accompanist.placeholder.material.shimmer
 import com.mxalbert.sharedelements.SharedElement
 import com.thekeeperofpie.artistalleydatabase.anime.LocalAnimeComposeSettings
 import com.thekeeperofpie.artistalleydatabase.anime.R
 import com.thekeeperofpie.artistalleydatabase.compose.conditionally
+
+@Composable
+fun MediaCoverImage(
+    screenKey: String,
+    mediaId: String?,
+    image: String?,
+    contentScale: ContentScale = ContentScale.Fit,
+    onSuccess: ((AsyncImagePainter.State.Success) -> Unit)? = null,
+    onError: ((AsyncImagePainter.State.Error) -> Unit)? = null,
+    modifier: Modifier = Modifier,
+) {
+    SharedElement(
+        key = "anime_media_${mediaId}_image",
+        screenKey = screenKey,
+    ) {
+        AsyncImage(
+            model = image,
+            contentScale = contentScale,
+            fallback = rememberVectorPainter(Icons.Filled.ImageNotSupported),
+            onSuccess = onSuccess,
+            onError = onError,
+            contentDescription = stringResource(R.string.anime_media_cover_image_content_description),
+            modifier = modifier.blurForScreenshotMode(),
+        )
+    }
+}
+
+@Composable
+fun MediaCoverImageNoSharedElement(
+    image: String?,
+    contentScale: ContentScale = ContentScale.Fit,
+    onSuccess: ((AsyncImagePainter.State.Success) -> Unit)? = null,
+    onError: ((AsyncImagePainter.State.Error) -> Unit)? = null,
+    modifier: Modifier = Modifier,
+) {
+    AsyncImage(
+        model = image,
+        contentScale = contentScale,
+        fallback = rememberVectorPainter(Icons.Filled.ImageNotSupported),
+        onSuccess = onSuccess,
+        onError = onError,
+        contentDescription = stringResource(R.string.anime_media_cover_image_content_description),
+        modifier = modifier.blurForScreenshotMode(),
+    )
+}
 
 @Composable
 fun MediaCoverImage(